Set the estimated runtime for service jobs to 1 day
This avoids deadlocks caused by under-provisioning workers to handle
services. If the estimated runtime of a service was short enough,
multiple services could be packed one after another into a single
"slice" of a reservation within the autoscaler. If other jobs were
waiting for those services to be established, or worse, if *all*
services needed to be established before real work could commence, the
result would be a bottleneck, or in the worst case a deadlock.

This treats services specially, forcing the autoscaler to always
provision enough nodes to handle *all* queued services to attempt to
avoid deadlocks.